Pagination

+++ backend/src/controllers/company/company.ts
@@ -21,10 +21,20 @@ companyRouter.get('/:id', async (req:express.Request, res:express.Response,next)
 })
 
 //returns all companies
-companyRouter.get('/', async (req:express.Request, res:express.Response)=>{
+companyRouter.get('/', async (req:any, res:express.Response)=>{
     let query = req.query
+    
     try{
         let companies
+        if('skip' in query && 'size'){
+            companies = await companyHelpers.findFromAndLimit(parseInt(query.skip) ,parseInt(query.size))
+            return res.status(200).send(JSON.stringify(companies))
+        }
+        if('top' in query){
+            companies = await companyHelpers.findTopCompaniesBasedOnAverage(parseInt(query.top))
+            return res.status(200).send(JSON.stringify(companies))
+        }
+        
         if (_.has(query,'name')){
             companies = await companyHelpers.searchByName(query.name)
             return res.status(200).send(JSON.stringify(companies))
@@ -51,41 +61,42 @@ companyRouter.get('/', async (req:express.Request, res:express.Response)=>{
 })
 
 //deletes a company given a id 
-companyRouter.delete('/:id', async(req:express.Request, res: express.Response)=>{
+companyRouter.delete('/:id', async function(req:any, res: express.Response){
     try{
-    let deletedComp = await Company.findByIdAndRemove(req.params.id)
-    if(!deletedComp){
-        let error:Error = new Error('The companany does not exist')
-        throw error
-    }
-    res.status(200)
-    res.send(JSON.stringify({message:`company ${req.params.id} is now deleted`}))
-   return
+        if (!req.authed.isAuthed){
+            return res.status(401).send()
+        }
+        let deletedComp = await Company.findByIdAndRemove(req.params.id)
+        if(!deletedComp){
+            let error:Error = new Error('The companany does not exist')
+            throw error
+        }
+        return  res.status(200).send(JSON.stringify({message:`company ${req.params.id} is now deleted`}))
     }catch(e){
         res.status(400)
-        res.send(JSON.stringify(e.message))
-        return
+        return res.send(JSON.stringify(e.message))
     }
 })
 
 //creates a new company
-companyRouter.post('/', async(req:express.Request, res: express.Response)=>{
-    let company = new Company({name:req.body.name})
+companyRouter.post('/', async(req:any, res: express.Response)=>{
+    
     try{
-        let existingCompany = await Company.findOne({name:req.body.name})
+        if (!req.authed.isAuthed){
+            return res.status(401).send()
+        }
+        let company = new Company({name:req.body.name})
+        let existingCompany = await Company.findOne({name:{
+            $regex: new RegExp(req.body.name,'i')}})
         if(existingCompany){
             let error:Error = new Error('company allready exist')
             throw error
         }
+        await company.save()
+        return res.status(200).send(JSON.stringify(company))
+
     }catch(error){
        return res.status(400).send(JSON.stringify(error.message))
-    }
-    try{
-        await company.save()
-       return res.status(200).send(JSON.stringify(company))
-    }catch(e){
-    
-        return res.status(400).send(JSON.stringify(e))
     }
 })
 export default companyRouter
